+++ R/process_trim_prevalence.R
@@ -12,11 +12,16 @@
 #' default the top level rank of the `ps`. taxonomic level(Kingdom, Phylum,
 #' Class, Order, Family, Genus, Species, Strains; default: NULL).
 #' @param cutoff (Optional). Numeric. the Prevalence threshold (default: 0.1).
+#' @param group (Optional). character. filtering features or samples by group
+#' (default: NULL).
 #' @param trim (Optional). Character. trimming to apply, the options include:
 #' * "none", return the original data without any actions.
 #' * "both", prevalence of features and samples more than cutoff.
 #' * "feature", prevalence of features more than cutoff.
+#' * "feature_group", prevalence of features more than cutoff by groups.
 #' * "sample", prevalence of samples more than cutoff.
+#' @param at_least_one (Optional). Logical. prevalence of at least one group
+#' meets cutoff (FALSE means all groups meet cutoff, default: FALSE).
 #'
 #' @return
 #'  A trimed `object` whose prevalence of features or samples more than cutoff.
@@ -34,7 +39,12 @@
 #'            "Order", "Family", "Genus",
 #'            "Species", "Strain", "unique"),
 #'     cutoff = 0.1,
-#'     trim = c("none", "both", "feature", "sample"))
+#'     group = NULL,
+#'     trim = c("none",
+#'       "both",
+#'       "feature", "feature_group",
+#'       "sample"),
+#'     at_least_one = FALSE)
 #'
 #' @examples
 #' \donttest{
@@ -59,22 +69,53 @@ trim_prevalence <- function(
     object,
     level = NULL,
     cutoff = 0.1,
-    trim = c("none", "both", "feature", "sample")){
+    group = NULL,
+    trim = c("none",
+             "both",
+             "feature", "feature_group",
+             "sample"),
+    at_least_one = FALSE) {
 
   # data(Zeybel_2022_gut)
   # object = Zeybel_2022_gut
-  # trim = "feature"
+  # trim = "sample_group"
   # level = "Genus"
   # cutoff = 0.1
+  # group = "LiverFatClass"
+  # at_least_one = FALSE
 
-  # data(Zeybel_2022_gut)
+  # data(Zeybel_2022_protein)
   # object = Zeybel_2022_protein
-  # trim = "both"
+  # trim = "feature"
   # level = NULL
-  # cutoff = 0.1
+  # cutoff = 0.95
+  # group = NULL
+  # at_least_one = FALSE
 
-  trim <- match.arg(trim, c("none", "both", "feature", "sample"))
+  # trim type
+  trim <- match.arg(trim, c("none",
+                            "both",
+                            "feature", "feature_group",
+                            "sample"))
+  if (length(grep("group", trim)) > 0) {
+    if (is.null(group)) {
+      stop("please provide `group` for filtering samples or features by group")
+    } else if (any(inherits(object, "phyloseq"),
+                   inherits(object, "SummarizedExperiment"))) {
+      if (inherits(object, "phyloseq")) {
+        metadata <- object@sam_data %>%
+          data.frame()
+      } else if (inherits(object, "SummarizedExperiment")) {
+        metadata <- object@colData %>%
+          data.frame()
+      }
+    } else {
+      stop("please using phyloseq or SummarizedExperiment object with
+           `group` in metadata for filtering samples or features by group")
+    }
+  }
 
+  # profile
   if (inherits(object, "phyloseq")) {
 
     # taxa level
@@ -89,29 +130,36 @@ trim_prevalence <- function(
     prf <- as(object$.Data, "matrix")
   } else if (inherits(object, "SummarizedExperiment")) {
     prf <- SummarizedExperiment::assay(object) %>%
-      data.frame() %>%
-      t()
+      data.frame() #%>%
+      #t()
   } else {
     prf <- object
   }
 
-  if (trim == "feature") {
+  # trimming
+  if (trim == "both") {
+    tmp1 <- trim_FeatureOrSample(prf, 1, cutoff)
+    tmp2 <- trim_FeatureOrSample(prf, 2, cutoff)
+    remain_features <- rownames(tmp1)
+    remain_samples <- rownames(tmp2)
+  } else if (trim == "feature") {
     tmp1 <- trim_FeatureOrSample(prf, 1, cutoff)
     remain_features <- rownames(tmp1)
     remain_samples <- colnames(prf)
-  } else if (trim == "sample") {
+  } else if(trim == "feature_group") {
+    tmp1 <- trim_FeatureByGroup(prf, cutoff,
+                                group, metadata, at_least_one)
+    remain_features <- rownames(tmp1)
+    remain_samples <- colnames(prf)
+  } else if(trim == "sample") {
     tmp2 <- trim_FeatureOrSample(prf, 2, cutoff)
     remain_features <- rownames(prf)
-    remain_samples <- rownames(tmp2)
-  } else if(trim == "both") {
-    tmp1 <- trim_FeatureOrSample(prf, 1, cutoff)
-    tmp2 <- trim_FeatureOrSample(prf, 2, cutoff)
-    remain_features <- rownames(tmp1)
     remain_samples <- rownames(tmp2)
   } else if(trim == "none") {
     return(object)
   }
 
+  # remain profile
   if (length(remain_features) > 1 & length(remain_samples) > 1) {
     prf_remain <- prf[remain_features, remain_samples]
   } else if (length(remain_features) > 1 & length(remain_samples) == 1) {
@@ -137,7 +185,7 @@ trim_prevalence <- function(
                                                    taxa_are_rows = taxa_are_rows(ps))
     object <- ps
   } else if (inherits(object, "SummarizedExperiment")) {
-    SummarizedExperiment::assay(object) <- t(prf_remain)
+    SummarizedExperiment::assay(object, withDimnames = TRUE) <- prf_remain #t(prf_remain)
   } else if (inherits(object, "environment")) {
     object <- phyloseq::otu_table(prf_remain, taxa_are_rows = taxa_are_rows(object))
   } else {
@@ -154,16 +202,69 @@ trim_FeatureOrSample <- function(x, nRow, threshold) {
   df_occ <- apply(x, nRow, function(x) {
     length(x[c(which(!is.na(x) & x!=0))]) / length(x)
   }) %>%
-    data.frame() %>% stats::setNames("Occ") %>%
-    tibble::rownames_to_column("type")
-  if(nRow == 1){
+    data.frame() %>% stats::setNames("Occ") #%>%
+    #tibble::rownames_to_column("type")
+  if (nRow == 1) {
     rownames(df_occ) <- rownames(x)
-  }else{
+  } else {
     rownames(df_occ) <- colnames(x)
   }
+
   df_KEEP <- apply(df_occ > threshold, 1, all) %>%
-    data.frame() %>% stats::setNames("Status") %>%
+    data.frame() %>%
+    stats::setNames("Status") %>%
     dplyr::filter(Status)
+
+  return(df_KEEP)
+}
+
+
+# the data is trimmed by threshold and group
+#' @keywords internal
+trim_FeatureByGroup <- function(x, threshold,
+                              group, metadata, at_least_one) {
+
+  # x = prf
+  # threshold = cutoff
+  # group = group
+  # metadata = metadata
+  # at_least_one = at_least_one
+
+  # groups for filtering
+  colnames(metadata)[which(colnames(metadata) == group)] <- "TempGroupName"
+  group_names <- unique(metadata$TempGroupName)
+  group_list <- lapply(group_names, function(x){
+    index <- which(metadata$TempGroupName == x)
+    group_samples <- rownames(metadata)[index]
+    return(group_samples)
+  })
+
+  # filter by feature
+  df_filter <- sapply(1:length(group_list), function(j) {
+      df_sub <- x[, colnames(x) %in% group_list[[j]]]
+      df_sub_matrix <- as.matrix(df_sub)
+      df_occ <- apply(df_sub_matrix, 1, function(x) {
+        length(x[c(which(!is.na(x) & x!=0))]) / length(x)
+      })
+      return(df_occ)
+    }) %>%
+      data.frame() %>%
+      stats::setNames(group_names) #%>%
+      #tibble::rownames_to_column("type")
+
+  #rownames(df_filter) <- rownames(x)
+
+  if (at_least_one) {
+    df_KEEP <- apply(df_filter, 1, function(x) {any(x > threshold)}) %>%
+      data.frame() %>%
+      stats::setNames("Status") %>%
+      dplyr::filter(Status)
+  } else {
+    df_KEEP <- apply(df_filter, 1, function(x) {all(x > threshold)}) %>%
+      data.frame() %>%
+      stats::setNames("Status") %>%
+      dplyr::filter(Status)
+  }
 
   return(df_KEEP)
 }